+++ src/main/kotlin/com/github/strindberg/emacsj/actions/search/ISearchMarkAction.kt
@@ -1,0 +1,8 @@
+package com.github.strindberg.emacsj.actions.search
+
+import com.github.strindberg.emacsj.search.ISearchMarkHandler
+import com.intellij.openapi.editor.actionSystem.EditorAction
+
+class ISearchMarkAction :
+    EditorAction(ISearchMarkHandler()),
+    ISearchAction