+++ src/main/kotlin/com/github/strindberg/emacsj/actions/search/ISearchSwapAction.kt
@@ -1,0 +1,8 @@
+package com.github.strindberg.emacsj.actions.search
+
+import com.github.strindberg.emacsj.search.ISearchSwapHandler
+import com.intellij.openapi.editor.actionSystem.EditorAction
+
+class ISearchSwapAction :
+    EditorAction(ISearchSwapHandler()),
+    ISearchAction