+++ CHANGELOG.md
@@ -4,9 +4,14 @@
 
 ## [Unreleased]
 
+### Added
+- Isearch can be finished with mark at start of final match.
+- Isearch can be finished with the final match selected.
+
 ### Fixed
 - Reversing search direction with empty search string does not trigger retrieval of previous search.
 - Updating plugin will less often require a restart.
+- Multi-caret search can now be reversed after failed match.
 
 ## [1.3.3] - 2025-01-27